refactor to modules
implemented a little erroro handling
